copy-to-clipboard fix and some rerender issues
